+++ YogaKit/Tests/YogaKitTests.m
@@ -153,6 +153,24 @@
   XCTAssertEqual(25, view2.frame.origin.y);
 }
 
+- (void)testMarkingDirtyOnlyWorksOnLeafNodes
+{
+  UIView *container = [[UIView alloc] initWithFrame:CGRectZero];
+  container.yoga.isEnabled = YES;
+
+  UIView *subview = [[UIView alloc] initWithFrame:CGRectZero];
+  subview.yoga.isEnabled = YES;
+  [container addSubview:subview];
+
+  XCTAssertFalse(container.yoga.isDirty);
+  [container.yoga markDirty];
+  XCTAssertFalse(container.yoga.isDirty);
+
+  XCTAssertFalse(subview.yoga.isDirty);
+  [subview.yoga markDirty];
+  XCTAssertTrue(subview.yoga.isDirty);
+}
+
 - (void)testThatMarkingLeafsAsDirtyWillTriggerASizeRecalculation
 {
   UIView *container = [[UIView alloc] initWithFrame:CGRectMake(0, 0, 500, 50)];